+++ src/repair/repair.controller.ts
@@ -8,6 +8,7 @@ import { RepairDeleteDto } from './dtos/repair.delete.dto';
 import { RepairUpdateDto } from './dtos/repair.update.dto';
 import { RepairQueryDto } from './dtos/repair.query.dto';
 import { MsgConst } from 'src/.const/msg.const';
+import { ObjectTool } from 'src/.tools/object.tool';
 
 @Controller('repair')
 export class RepairController {
@@ -16,6 +17,7 @@ export class RepairController {
   @Post('create')
   async create(@Body() repairCreateDto: RepairCreateDto) {
     if (!NumberTool.isInteger(repairCreateDto.checkingUid)) return Result.fail(MsgConst.powerLowE);
+    if (!ObjectTool.isBodyExist(repairCreateDto)) return Result.fail(MsgConst.bodyNotExistE);
     if (!StringTool.isLengthInRange(repairCreateDto.body.title, 0, 20)) return Result.fail(MsgConst.titleLengthE);
     if (!StringTool.isLengthInRange(repairCreateDto.body.content, 0, 255)) return Result.fail(MsgConst.contentLengthE);
     if (!StringTool.isLengthInRange(repairCreateDto.body.address, 0, 50)) return Result.fail(MsgConst.addressLengthE);
@@ -26,6 +28,7 @@ export class RepairController {
   @Post('delete')
   async delete(@Body() repairDeleteDto: RepairDeleteDto) {
     if (!NumberTool.isInteger(repairDeleteDto.checkingUid)) return Result.fail(MsgConst.powerLowE);
+    if (!ObjectTool.isBodyExist(repairDeleteDto)) return Result.fail(MsgConst.bodyNotExistE);
     if (!NumberTool.isInteger(repairDeleteDto.body.id)) return Result.fail(MsgConst.idNotExistE);
 
     return await this.repairService.delete(repairDeleteDto);
@@ -34,6 +37,7 @@ export class RepairController {
   @Post('update')
   async update(@Body() repairUpdateDto: RepairUpdateDto) { 
     if (!NumberTool.isInteger(repairUpdateDto.checkingUid)) return Result.fail(MsgConst.powerLowE);
+    if (!ObjectTool.isBodyExist(repairUpdateDto)) return Result.fail(MsgConst.bodyNotExistE);
     if (!NumberTool.isInteger(repairUpdateDto.body.id)) return Result.fail(MsgConst.idNotExistE);
     if (!NumberTool.isIntegerInRange(repairUpdateDto.body.status,2,3)) return Result.fail(MsgConst.statusE);
 
@@ -43,6 +47,7 @@ export class RepairController {
   @Post('query')
   async query(@Body() repairQueryDto: RepairQueryDto) { 
     if (!NumberTool.isInteger(repairQueryDto.checkingUid)) return Result.fail(MsgConst.powerLowE);
+    if (!ObjectTool.isBodyExist(repairQueryDto)) return Result.fail(MsgConst.bodyNotExistE);
     if (!NumberTool.isIntegerInRange(repairQueryDto.body.pageSize, 1, 100)) return Result.fail(MsgConst.pageSizeE);
     if (!NumberTool.isInteger(repairQueryDto.body.pageIndex)) return Result.fail(MsgConst.pageIndexE);
